Adds PCH

+++ Keg/src/FileSystem/FileSystem.cpp
@@ -1,5 +1,6 @@
+#include "stadx.h"
+
 #include "FileSystem.h"
-#include <filesystem>
 
 bool Keg::FileSystem::createFolder(std::string folderPath, std::string folderName)
 {
@@ -15,8 +16,8 @@ bool Keg::FileSystem::changeFolderName(std::string folderPath, std::string newNa
 {
 	try {
 		std::string temp = folderPath;
-		int index = folderPath.find_last_of("\\");
-		folderPath = folderPath.substr(0, int(index + 1));
+		size_t index = folderPath.find_last_of("\\");
+		folderPath = folderPath.substr(0, size_t(index + 1));
 		folderPath += newName;
 		std::filesystem::rename(temp, folderPath);
 		return true;
@@ -29,7 +30,6 @@ bool Keg::FileSystem::changeFolderName(std::string folderPath, std::string newNa
 bool Keg::FileSystem::copyFile(std::string filePath, std::string fileNewPath)
 {
 	try {
-		
 		std::filesystem::copy(filePath, fileNewPath);
 		return true;
 	}
